requirements.txt
PyQt5
pyobjc